qr-code-maker/requirements.txt
qrcode
Pillow
python-vlc